Makes the Plugin Blocks Report more generic
Instead of using a predefined list of shortcodes, the report now pulls in all defined shortcodes and filters them on being a shortcake shortcode.
This is usefull for the development of new blocks being directly in the report as well as for NRO added blocks being visible in the same view.